# docs/21-api-openapi.yaml
openapi: 3.1.0
info:
  title: Gym Management System API
  version: 1.0.0
  description: Machine-readable V1 contract. Base monetary unit is integer minor units (100 = 1 EGP). Arabic RTL owner UI.
servers:
- url: /api/v1
security:
- cookieAuth: []
...
paths:
...
  /plans:
    get:
      tags:
      - Plans
      parameters:
      - name: enabled
        in: query
        schema:
          type: boolean
      - $ref: '#/components/schemas/Page'
      - $ref: '#/components/schemas/Limit'
      responses:
        '200':
          description: نجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanListResponse'
        '400':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
      - Plans
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanMutationRequest'
      responses:
        '201':
          description: نجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'
        '400':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /plans/{planId}:
    get:
      tags:
      - Plans
      parameters:
      - $ref: '#/components/schemas/PlanId'
      responses:
        '200':
          description: نجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'
        '401':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    patch:
      tags:
      - Plans
      parameters:
      - $ref: '#/components/schemas/PlanId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePlanRequest'
      responses:
        '200':
          description: نجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'
        '400':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /plans/{planId}/enable:
    post:
      tags:
      - Plans
      parameters:
      - $ref: '#/components/schemas/PlanId'
      responses:
        '200':
          description: نجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'
        '401':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /plans/{planId}/disable:
    post:
      tags:
      - Plans
      parameters:
      - $ref: '#/components/schemas/PlanId'
      responses:
        '200':
          description: نجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'
        '401':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: gym_session
...
  schemas:
    ApiError:
      type: object
      additionalProperties: false
      required:
      - statusCode
      - code
      - message
      - details
      properties:
        statusCode:
          type: integer
        code:
          type: string
        message:
          type: string
        details:
          anyOf:
          - type: object
            additionalProperties: true
          - type: 'null'
    Pagination:
      type: object
      additionalProperties: false
      required:
      - page
      - limit
      - totalItems
      - totalPages
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
        totalItems:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
...
    PlanPrice:
      type: object
      additionalProperties: false
      required:
      - durationMonths
      - priceMinor
      properties:
        durationMonths:
          type: integer
          enum:
          - 1
          - 3
          - 6
          - 12
        priceMinor:
          type: integer
          minimum: 0
          description: Integer minor units; 100 = 1 EGP
    Plan:
      type: object
      additionalProperties: false
      required:
      - id
      - name
      - isEnabled
      - prices
      - createdAt
      - updatedAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        isEnabled:
          type: boolean
        prices:
          type: array
          minItems: 4
          maxItems: 4
          items:
            $ref: '#/components/schemas/PlanPrice'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PlanMutationRequest:
      type: object
      additionalProperties: false
      required:
      - name
      - prices
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        prices:
          type: array
          minItems: 4
          maxItems: 4
          items:
            $ref: '#/components/schemas/PlanPrice'
    UpdatePlanRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        prices:
          type: array
          minItems: 4
          maxItems: 4
          items:
            $ref: '#/components/schemas/PlanPrice'
...
    PlanListResponse:
      type: object
      additionalProperties: false
      required:
      - items
      - pagination
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Plan'
        pagination:
          $ref: '#/components/schemas/Pagination'
...